requirements.txt
Flask~=3.0.2
stcp_api~=1.1.0
